BulletproofTxManager
See this ticket: https://www.pivotaltracker.com/story/show/171753295
See design proposal: https://www.notion.so/chainlink/DESIGN-PROPOSAL-BulletproofTxManager-81436286a142411d95e336b0c295c457
See design doc: https://www.notion.so/chainlink/BulletproofTxManager-Architecture-Overview-9dc62450cd7a443ba9e7dceffa1a8d6b

This commit changes ethtx so that it writes to the `eth_txes`
table instead when ENABLE_BULLETPROOF_TX_MANAGER flag is turned on, and
introduces a BulletproofTxManager services to ensure the transaction is
sent.

BulletproofTxManager is a complete replacement for the legacy TxManager and doesn't use any of the original code. It features the following improvements:

1. Safe to run concurrently across multiple nodes (horizontally scalable)
2. Clear concurrency model (no race conditions)
3. Error handling is all done in one place and is clearer and easier to reason about
4. Clear separation of responsibilities and more comprehensible code
5. Guarantees around nonce tracking that **ensure** that there can never be gaps in the nonce sequence
6. Decoupling of `ethtx` and `txmanager` makes `ethtx` much smaller and simpler
7. Properly supports multiple private keys (node operators can now specify `fromAddress` in ethtx to pin a job to a particular address)
8. It is no longer required to ask for receipts on every new block, resulting in far fewer requests to the ethereum node (by an order of magnitude or more)
9. Deeper re-org protection (on the order of hundreds of blocks is possible)
10. Predictable behaviour in the event that an external account uses the nonce
11. Clear guarantees and design ensure transactions will ALWAYS be put on the longest canonical chain within well defined constraints.
12. Any new code can now easily send transactions simply by writing to the `eth_txes` table and letting BulletproofTxManager do the rest.

A new setting `ETH_FINALITY_DEPTH` is exposed which allows a node operator to specify how many blocks deep they want re-org protection for. This defaults to 50.

BulletproofTxManager is enabled by setting `ENABLE_BULLETPROOF_TX_MANAGER=true`. This will trigger an automatic migration to set the initial nonce correctly. It is important to note that THIS IS A ONE-WAY MIGRATION. Once it has been turned on once, it may not ever be turned off. This is by design, since the BulletproofTxManager architecture absolutely requires owning the nonce for an account (all those safety guarantees come with a cost) so allowing the old TxManager to mess with it would lead to undefined behaviour.

An escape hatch is provided with a new client command `setnextnonce`. This allows a user to manually set the next nonce in the sequence for a given account. It should never be necessary and is included "just in case".